+++ CHANGELOG.md
@@ -20,3 +20,4 @@
 - Create login with jwt
 - Create API for Waste Type
 - Create API for Deposit
+- Create API for withdrawal